requirements.txt
transformers==4.54.1
sentence-transformers==5.0.0
redis==6.4.0
redisvl==0.8.2
python-dotenv==1.1.1

langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.8
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.2
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.2
langgraph-sdk==0.2.0
langsmith==0.4.9

panel==1.7.5
pandas==2.2.3
jaro-winkler==2.0.3
seaborn==0.13.2
fuzzywuzzy==0.18.0

gradio==5.49.1
gradio-client==1.13.3
tavily-python==0.7.12
